+++ src/main/java/roomescape/auth/AdminInterceptor.java
@@ -1,6 +1,5 @@
 package roomescape.auth;
 
-import io.jsonwebtoken.JwtException;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import org.springframework.stereotype.Component;
