+++ src/main/java/roomescape/auth/AuthConstants.java
@@ -1,0 +1,5 @@
+package roomescape.auth;
+
+public class AuthConstants {
+    public static final String AUTH_TOKEN_COOKIE = "token";
+}
