+++ src/main/java/roomescape/member/LoginRequest.java
@@ -1,4 +1,14 @@
 package roomescape.member;
 
 public class LoginRequest {
+    private String email;
+    private String password;
+
+    public String getEmail() {
+        return email;
+    }
+
+    public String getPassword() {
+        return password;
+    }
 }
